+++ forms/src/forms.ts
@@ -4,6 +4,7 @@ export * from './directives/numberInput/numberInputValidator.directive';
 export * from './directives/numberInput/numberInputControlValueAccessor.directive';
 export * from './directives/hasError/hasError.directive';
 export * from './directives/groupHasError/groupHasError.directive';
+export * from './directives/requiredClass/requiredClass.directive';
 export * from './services/submitted';
 export * from './misc/validators';
 export * from './misc/utils';